[INJIMOB-2761]: update the readme with latest changes and move the jwt exception cladd to correct folder (#31)
